backendpy/requirements.txt
dnspython==2.8.0
mysql-connector-python==9.5.0
pymongo==4.15.5
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
